rework routes
